Restructure Flask app with routes and templates
