+++ README.md
@@ -1,1 +1,3 @@
 # DataWasher
+
+A simple Flask application skeleton.
