+++ requirements.txt
@@ -1,0 +1,1 @@
+Flask==3.1.1
