+++ templates/greeting.html
@@ -1,0 +1,7 @@
+<!doctype html>
+<html>
+<head><title>Greeting</title></head>
+<body>
+    <h1>Welcome!</h1>
+</body>
+</html>
