+++ templates/home.html
@@ -1,0 +1,7 @@
+<!doctype html>
+<html>
+<head><title>Home</title></head>
+<body>
+    <h1>Home Page</h1>
+</body>
+</html>
