+++ templates/login.html
@@ -1,0 +1,7 @@
+<!doctype html>
+<html>
+<head><title>Login</title></head>
+<body>
+    <h1>Login Page</h1>
+</body>
+</html>
